Support: Adding OwnPtr & NonNullOwnPtr
Adding own unique ptr which has a non-null version
Changing raw pointer to (NonNull)-OwnPtr

FIXME: Don't include RenderContext into EngineLoop to not leak headers

+++ Source/Engine/include/HyperEngine/Support/ErrorHandling.hpp
@@ -9,13 +9,13 @@
 #include "HyperEngine/Support/Compiler.hpp"
 #include "HyperEngine/Support/Prerequisites.hpp"
 
-#include <string_view>
+#include <cstddef>
 
 namespace HyperEngine::Detail
 {
 	[[noreturn]] void unreachable(
 		const char *file_name = __builtin_FILE(),
-		uint32_t line = __builtin_LINE());
+		size_t line = __builtin_LINE());
 } // namespace HyperEngine::Detail
 
 #if HYPERENGINE_DEBUG
